:art: 课表 item 层统一在 course:view 模块中，并提供 PlatformCourse*Item 下层每个平台的配置
