+++ cyxbs-pages/course/src/commonMain/kotlin/com/cyxbs/pages/course/frame/decoration/LinkLessonDecorationViewModel.kt
@@ -4,14 +4,14 @@ import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewModelScope
 import com.cyxbs.components.base.ui.BaseViewModel
 import com.cyxbs.pages.course.api.LessonByWeeks
-import com.cyxbs.pages.course.view.frame.item.LinkLessonItem
-import com.cyxbs.pages.course.view.frame.item.LinkLessonItemFactory
 import com.cyxbs.pages.course.model.LessonRepository
 import com.cyxbs.pages.course.model.LinkLessonRepository
 import com.cyxbs.pages.course.view.decoration.CoursePageDecoration
 import com.cyxbs.pages.course.view.item.CourseItemHierarchy
 import com.cyxbs.pages.course.view.item.CourseItemWhatTime
 import com.cyxbs.pages.course.view.item.ItemHierarchyWhatTime
+import com.cyxbs.pages.course.view.item.impl.CourseLinkLessonItem
+import com.cyxbs.pages.course.view.item.impl.PlatformCourseLinkLessonItemFactory
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -29,7 +29,8 @@ import kotlinx.coroutines.flow.onEach
  * @date 2025/10/18
  */
 class LinkLessonDecorationViewModel(
-  val hierarchy: CourseItemHierarchy<LinkLessonItem>,
+  val hierarchy: CourseItemHierarchy<CourseLinkLessonItem>,
+  val platformItemFactory: PlatformCourseLinkLessonItemFactory,
 ) : BaseViewModel(), CoursePageDecoration {
 
   init {
@@ -60,10 +61,10 @@ class LinkLessonDecorationViewModel(
     hierarchy.reset(buildList {
       data.forEach { lesson ->
         // 添加进整学期
-        add(LinkLessonWhatTime(0, lesson))
+        add(LinkLessonWhatTime(0, lesson, platformItemFactory))
         // 添加进每周
         lesson.week.forEach { week ->
-          add(LinkLessonWhatTime(week, lesson))
+          add(LinkLessonWhatTime(week, lesson, platformItemFactory))
         }
       }
     })
@@ -78,7 +79,8 @@ class LinkLessonDecorationViewModel(
 private data class LinkLessonWhatTime(
   val page: Int,
   val lesson: LessonByWeeks,
-) : ItemHierarchyWhatTime<LinkLessonItem>() {
+  val platformItemFactory: PlatformCourseLinkLessonItemFactory,
+) : ItemHierarchyWhatTime<CourseLinkLessonItem>() {
   override val now: MutableStateFlow<CourseItemWhatTime.Fixed> = MutableStateFlow(
     CourseItemWhatTime.Fixed(
       page = page,
@@ -88,15 +90,16 @@ private data class LinkLessonWhatTime(
     )
   )
 
-  override fun createItem(coroutineScope: CoroutineScope): LinkLessonItem {
-    return LinkLessonItemFactory.get().createLinkLessonItem(
+  override fun createItem(coroutineScope: CoroutineScope): CourseLinkLessonItem {
+    return CourseLinkLessonItem(
       whatTime = this,
       coroutineScope = coroutineScope,
       lesson = lesson,
+      platformItemFactory = platformItemFactory,
     )
   }
 
-  override fun compareTo(other: ItemHierarchyWhatTime<LinkLessonItem>): Int {
+  override fun compareTo(other: ItemHierarchyWhatTime<CourseLinkLessonItem>): Int {
     return 0.compareBy(other) {
       -it.now.value.page // page 越小越在上
     }.compareBy(other) {
@@ -111,8 +114,8 @@ private data class LinkLessonWhatTime(
   }
 
   private inline fun Int.compareBy(
-    other: ItemHierarchyWhatTime<LinkLessonItem>,
-    compare: (ItemHierarchyWhatTime<LinkLessonItem>) -> Int
+    other: ItemHierarchyWhatTime<CourseLinkLessonItem>,
+    compare: (ItemHierarchyWhatTime<CourseLinkLessonItem>) -> Int
   ): Int {
     if (this != 0) return this
     val a = compare.invoke(this@LinkLessonWhatTime)
